+++ functions/Widgets.fnc.php
@@ -1081,9 +1081,9 @@ function Widgets( $item, &$myextra = null )
 			}
 
 			$extra['search'] .= '<tr class="st"><td>' . _( 'Balance' ) . '</td><td><label>' . _( 'Between' ) .
-			' <input type="text" name="balance_low" size="5" maxlength="10" /></label>' .
+			' <input type="number" name="balance_low" step="any" /></label>' .
 			' <label>&amp;' .
-			' <input type="text" name="balance_high" size="5" maxlength="10" /></label>
+			' <input type="number" name="balance_high" step="any" /></label>
 			</td></tr>';
 
 		break;